change js build
