--- javascript-modules-engine/src/javascript/i18next/init.ts
@@ -1,18 +1,0 @@
-import i18n from "i18next";
-import backend from "./backend";
-import { initReactI18next } from "react-i18next";
-
-export default () => {
-  i18n
-    .use(backend)
-    .use(initReactI18next)
-    .init({
-      fallbackLng: "en",
-      ns: "javascript-modules-engine",
-      defaultNS: "javascript-modules-engine",
-      initImmediate: false,
-      react: {
-        useSuspense: false,
-      },
-    });
-};
